requirements.txt
streamlit == 1.30.0
python-dotenv == 1.0.1
python == 3.10.6
google-generativeai==0.8.3